+++ ruoyi-modules/ruoyi-flower/src/main/java/org/dromara/flower/domain/bo/FolwerDeliveryPriceBo.java
@@ -9,6 +9,8 @@ import lombok.Data;
 import lombok.EqualsAndHashCode;
 import jakarta.validation.constraints.*;
 
+import java.math.BigDecimal;
+
 /**
  * 物流计费业务对象 folwer_delivery_price
  *
@@ -64,27 +66,27 @@ public class FolwerDeliveryPriceBo extends BaseEntity {
     /**
      * 计费首重
      */
-    private Long firstWeight;
+    private Double firstWeight;
 
     /**
      * 续重重量
      */
-    private Long additionalWeight;
+    private Double additionalWeight;
 
     /**
      * 首重价格
      */
-    private Long firstWeightPrice;
+    private BigDecimal firstWeightPrice;
 
     /**
      * 续重1价格
      */
-    private Long additionalWeightPrice;
+    private BigDecimal additionalWeightPrice;
 
     /**
      * 续重2价格
      */
-    private Long additionalWeightPricel;
+    private BigDecimal additionalWeightPricel;
 
     /**
      * 是否使用，1:使用，0:不使用
