+++ ruoyi-modules/ruoyi-flower/src/main/java/org/dromara/flower/domain/vo/FolwerDeliveryPriceVo.java
@@ -10,6 +10,7 @@ import lombok.Data;
 
 import java.io.Serial;
 import java.io.Serializable;
+import java.math.BigDecimal;
 import java.util.Date;
 
 
@@ -80,31 +81,31 @@ public class FolwerDeliveryPriceVo implements Serializable {
      * 计费首重
      */
     @ExcelProperty(value = "计费首重")
-    private Long firstWeight;
+    private Double firstWeight;
 
     /**
      * 续重重量
      */
     @ExcelProperty(value = "续重重量")
-    private Long additionalWeight;
+    private Double additionalWeight;
 
     /**
      * 首重价格
      */
     @ExcelProperty(value = "首重价格")
-    private Long firstWeightPrice;
+    private BigDecimal firstWeightPrice;
 
     /**
      * 续重1价格
      */
     @ExcelProperty(value = "续重1价格")
-    private Long additionalWeightPrice;
+    private BigDecimal additionalWeightPrice;
 
     /**
      * 续重2价格
      */
     @ExcelProperty(value = "续重2价格")
-    private Long additionalWeightPricel;
+    private BigDecimal additionalWeightPricel;
 
     /**
      * 是否使用，1:使用，0:不使用
